% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_hydro_meta.R
\name{dd_hydro_meta}
\alias{dd_hydro_meta}
\title{dd_hydro_meta}
\usage{
dd_hydro_meta(prov_terr)
}
\arguments{
\item{prov_terr}{Province/territory short code options are \describe{
  \item{NB}{New Brunswick}
  \item{PE}{Prince Edward Island}
  \item{NS}{Nova Scotia}
  \item{ON}{Ontario}
  \item{QC}{Quebec}
  \item{NL}{Newfoundland}
  \item{MB}{Manitoba}
  \item{AB}{Alberta}
  \item{SK}{Saskatchewan}
  \item{NU}{Nunavut}
  \item{NT}{Northwest Territories}
  \item{BC}{British Columbia}
  \item{YT}{Yukon}
}}
}
\description{
Get metadata for available stations, filter by province or territory.
#' See \href{https://dd.weather.gc.ca/hydrometric/doc/hydrometric_README.txt}{DataMart documentation} for more info.
}
\examples{
dd_hydro_meta(prov_terr = "ON")

}
