% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dd_to_zrxp.R
\name{dd_to_zrxp}
\alias{dd_to_zrxp}
\title{dd_to_zrxp}
\usage{
dd_to_zrxp(stn_data, stn_import_num, min_timestamp)
}
\arguments{
\item{stn_data}{The time series data being converted. Should be two columns \describe{
 \item{First column}{Timestamp}
 \item{Second column}{Value}
}}

\item{stn_import_num}{The WISKI import number for time series data being converted.}

\item{min_timestamp}{POSIXct format timestamp. The minimum timestamp to filter for before exporting.}
}
\value{
Vector of lines to write to .zrxp file
}
\description{
Convert data file to zrxp for WISKI ingestion.
}
